+++ tools/ui/src/lib/constants/index.ts
@@ -38,6 +38,8 @@ export * from './precision';
 export * from './processing-info';
 export * from './routes';
 export * from './sandbox';
+export * from './sandbox-harness';
+export * from './sandbox-worker';
 export * from './settings-keys';
 export * from './settings-registry';
 export * from './supported-file-types';
