+++ tools/ui/src/lib/constants/sandbox.ts
@@ -1,4 +1,4 @@
-import { ToolCallType } from '$lib/enums';
+import { JsonSchemaType, ToolCallType } from '$lib/enums';
 import type { OpenAIToolDefinition } from '$lib/types';
 
 export const SANDBOX_TOOL_NAME = 'run_javascript';
@@ -9,6 +9,10 @@ export const SANDBOX_TIMEOUT_MS_MAX = 30000;
 
 export const SANDBOX_OUTPUT_MAX_CHARS = 8192;
 
+export const SANDBOX_EMPTY_OUTPUT = '(no output)';
+
+export const SANDBOX_TRUNCATION_NOTICE = '[output truncated]';
+
 export const SANDBOX_TOOL_DEFINITION: OpenAIToolDefinition = {
 	type: ToolCallType.FUNCTION,
 	function: {
@@ -18,14 +22,14 @@ export const SANDBOX_TOOL_DEFINITION: OpenAIToolDefinition = {
 			'Top level await is supported. Use console.log to print intermediate values; ' +
 			'a top level return statement is captured as the result.',
 		parameters: {
-			type: 'object',
+			type: JsonSchemaType.OBJECT,
 			properties: {
 				code: {
-					type: 'string',
+					type: JsonSchemaType.STRING,
 					description: 'JavaScript source to execute'
 				},
 				timeout_ms: {
-					type: 'number',
+					type: JsonSchemaType.NUMBER,
 					description: `Execution timeout in milliseconds, default ${SANDBOX_TIMEOUT_MS_DEFAULT}, max ${SANDBOX_TIMEOUT_MS_MAX}`
 				}
 			},
